refactor: share tanstack route helpers
